feat(healthcheck): return availability in float instead of percentage
**Issue** [APIM-7353](https://gravitee.atlassian.net/browse/APIM-7353)

## Description

The API send availability with numbers between 0 and 1 with 4 digits of precision.

+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/main/java/io/gravitee/apim/core/api_health/use_case/AvailabilityUseCase.java
@@ -69,5 +69,5 @@ public class AvailabilityUseCase {
         }
     }
 
-    public record Output(int global, Map<String, Integer> byField) {}
+    public record Output(float global, Map<String, Float> byField) {}
 }
